Update v2.3
Tried to symmetrize the vertical and horizontal speeds but with no real success... Managed to make the vertical movement laggy (and I don't want that).  Also, fixed the teleporting in the opposite direction if the snake is moving in a direction (ex: IF the snake is moving right, you cannot go left).