Update 2.3.4
Started transfering the game into the graphics.h window. 
Got the borders of the map drawn , also a prototype for the score display.
We also went for a square matrix now , since in the graphics.h window a 40x40 window is a square , not a rectangle like the one in cmd.