Update v3.0
Implemented speed boost correctly (finally) using <time.h> functions. When collected, the speed increases for ~12 seconds.
Added pause /  resume system.
Score is now a multiple of 10 for a more "arcade" feel.
Sprites moved in a header.
There were problems when the game was restarted in the same compiling session... When the game is quit, you have to open it again to play.